AUREM-PROJECT/backend/requirements.txt
aiofile==3.9.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.13.0
apimatic-core==0.2.24
apimatic-core-interfaces==0.1.8
apimatic-requests-client-adapter==0.1.10
APScheduler==3.11.2
attrs==25.4.0
Authlib==1.6.8
backports.tarfile==1.2.0
bcrypt==4.1.3
beartype==0.22.9
black==25.12.0
boto3==1.42.16
botocore==1.42.16
build==1.4.2
CacheControl==0.14.4
cachetools==6.2.4
caio==0.9.25
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.5
click==8.3.1
cloudinary==1.44.1
cryptography==46.0.3
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cyclopts==4.10.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
durationpy==0.10
ecdsa==0.19.1
elevenlabs==2.36.1
email-validator==2.3.0
emergentintegrations==0.1.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.110.1
fastmcp==3.2.0
fastuuid==0.14.0
filelock==3.20.1
flake8==7.3.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.1
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gradio_client==2.4.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.3
http_ece==1.2.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jq==1.10.0
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
kubernetes==35.0.0
langdetect==1.0.9
librt==0.7.4
limits==5.8.0
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.1
more-itertools==10.8.0
motor==3.3.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.0
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
oauthlib==3.3.1
onnxruntime==1.24.4
openai==1.99.9
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.8
overrides==7.7.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathable==0.5.0
pathspec==0.12.1
paypal-checkout-serversdk==1.0.3
paypal-server-sdk==2.2.0
paypalhttp==1.0.1
paypalrestsdk==1.13.3
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
py-key-value-aio==0.4.4
py-vapid==1.9.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.14.0
pycountry==26.2.16
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pyOpenSSL==25.3.0
pyparsing==3.3.1
pyperclip==1.11.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
pywebpush==2.3.0
PyYAML==6.0.3
pyzbar==0.1.9
redis==7.4.0
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.19.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.1
s3transfer==0.16.0
s5cmd==0.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
SecretStorage==3.5.0
sendgrid==6.12.5
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sse-starlette==3.3.4
starlette==0.37.2
stripe==14.1.0
sympy==1.14.0
telnyx==4.90.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
triton==3.6.0
twilio==9.9.1
typer==0.20.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uncalled-for==0.2.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
wrapt==2.1.2
yarl==1.22.0
zipp==3.23.0
dateparser==1.4.0
